test2

+++ Tables/Personnes.sql
@@ -4,6 +4,6 @@ CREATE TABLE Personnes (
     id_personne VARCHAR(255) PRIMARY KEY NOT NULL,
     nom VARCHAR(255),
     prenom VARCHAR(255),
-    age INT(3),
+    age NUMBER(25),
     Adresseid_adresse VARCHAR(255)
 );
